# src/test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: api-datawarehousereader
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "5.0"
servers:
- url: http://localhost:50132
  description: Generated server url
...
paths:
...
  /{municipalityId}/invoices:
    get:
      tags:
      - Invoice
      summary: Get basic invoice information
      description: Resource returns all invoices matching provided search parameters
      operationId: getInvoices
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: searchParams
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/InvoiceParameters"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/invoices/{organizationNumber}/{invoiceNumber}/details:
    get:
      tags:
      - Invoice
      summary: Get invoice details for invoice matching issuer of provided organization
        number and having invoice number matching provided invoice number
      operationId: getInvoiceDetails
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: organizationNumber
        in: path
        description: Organization number of invoice issuer
        required: true
        schema:
          type: string
        example: 5565027223
      - name: invoiceNumber
        in: path
        description: Invoice number
        required: true
        schema:
          type: integer
          format: int64
        example: 805137494
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceDetail"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/customer/engagements:
    get:
      tags:
      - Customer
      summary: Get basic customer information
      description: Resource returns all customer engagements matching provided search
        parameters
      operationId: getCustomerEngagements
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
            description: PartyId (e.g. a personId or an organizationId)
            examples:
            - 81471222-5798-11e9-ae24-57fa13b361e1
      - name: customerNumber
        in: query
        description: Customer number
        required: false
        schema:
          type: string
          description: Customer number
          examples:
          - "10007"
      - name: organizationNumber
        in: query
        description: Organization number for counterpart of engagement
        required: false
        schema:
          type: string
          description: Organization number for counterpart of engagement
          examples:
          - "5565027223"
      - name: organizationName
        in: query
        description: Organization name for counterpart of engagement
        required: false
        schema:
          type: string
          description: Organization name for counterpart of engagement
          examples:
          - Sundsvall Elnät
      - name: sortBy
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
            description: The properties to sort on
            examples:
            - propertyName
      - name: sortDirection
        in: query
        description: The sort order direction
        required: false
        schema:
          $ref: "#/components/schemas/Direction"
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          description: Page number
          examples:
          - "1"
          minimum: 1
      - name: limit
        in: query
        description: Result size per page. Maximum allowed value is dynamically configured
        required: false
        schema:
          type: integer
          format: int32
          description: Result size per page. Maximum allowed value is dynamically
            configured
          examples:
          - "15"
          minimum: 1
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerEngagementResponse"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/customer/details:
    get:
      tags:
      - Customer
      summary: Get detailed customer information
      description: Resource returns all customer details matching provided search
        parameters
      operationId: getCustomerDetails
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
            description: PartyId (e.g. a personId or an organizationId)
            examples:
            - 81471222-5798-11e9-ae24-57fa13b361e1
      - name: customerEngagementOrgId
        in: query
        description: Organization id for customer engagements
        required: true
        schema:
          type: string
          description: Organization id for customer engagements
          examples:
          - "5565027225"
      - name: fromDateTime
        in: query
        description: Date and time for when to search for changes from. Format is
          yyyy-MM-dd'T'HH:mm:ss.SSSXXX. If omitted changes within the last year will
          be returned.
        required: false
        schema:
          type: string
          format: date-time
          description: Date and time for when to search for changes from. Format is
            yyyy-MM-dd'T'HH:mm:ss.SSSXXX. If omitted changes within the last year
            will be returned.
          examples:
          - 2000-10-31T01:30:00.000-05:00
      - name: sortBy
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
            description: The properties to sort on
            examples:
            - propertyName
      - name: sortDirection
        in: query
        description: The sort order direction
        required: false
        schema:
          $ref: "#/components/schemas/Direction"
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          description: Page number
          examples:
          - "1"
          minimum: 1
      - name: limit
        in: query
        description: Result size per page. Maximum allowed value is dynamically configured
        required: false
        schema:
          type: integer
          format: int32
          description: Result size per page. Maximum allowed value is dynamically
            configured
          examples:
          - "15"
          minimum: 1
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetailsResponse"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        status:
          $ref: "#/components/schemas/StatusType"
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties: {}
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties: {}
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
...
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    CustomerType:
      type: string
      description: Customer type
      enum:
      - Enterprise
      - Private
      examples:
      - Enterprise
    Direction:
      type: string
      enum:
      - ASC
      - DESC
    InvoiceParameters:
      type: object
      description: Invoice request parameters model
      properties:
        page:
          type: integer
          format: int32
          default: 1
          description: Page number
          examples:
          - "1"
          minimum: 1
        limit:
          type: integer
          format: int32
          description: Result size per page. Maximum allowed value is dynamically
            configured
          examples:
          - "15"
          minimum: 1
        sortBy:
          type: array
          items:
            type: string
            description: The properties to sort on
            examples:
            - propertyName
        sortDirection:
          $ref: "#/components/schemas/Direction"
          description: The sort order direction
          examples:
          - ASC
        customerNumber:
          type: array
          items:
            type: string
            description: Customer numbers
            examples:
            - "39195"
        customerType:
          $ref: "#/components/schemas/CustomerType"
        facilityIds:
          type: array
          items:
            type: string
            description: Facility ids
            examples:
            - "735999109151401011"
        invoiceNumber:
          type: integer
          format: int64
          description: Invoice number
          examples:
          - "767915994"
        invoiceDateFrom:
          type: string
          format: date
          description: Earliest invoice date. Format is YYYY-MM-DD.
          examples:
          - 2022-01-01
        invoiceDateTo:
          type: string
          format: date
          description: Latest invoice date. Format is YYYY-MM-DD.
          examples:
          - 2022-01-31
        invoiceName:
          type: string
          description: Invoice name
          examples:
          - 765801493.pdf
        invoiceType:
          type: string
          description: Invoice type
          examples:
          - Faktura
        invoiceStatus:
          type: string
          description: Invoice status
          examples:
          - Skickad
        ocrNumber:
          type: integer
          format: int64
          description: OCR number
          examples:
          - "767915994"
        dueDateFrom:
          type: string
          format: date
          description: Earliest due date. Format is YYYY-MM-DD.
          examples:
          - 2022-01-01
        dueDateTo:
          type: string
          format: date
          description: Latest due date. Format is YYYY-MM-DD.
          examples:
          - 2022-01-31
        organizationGroup:
          type: string
          description: Organization group
          examples:
          - stadsbacken
        organizationNumber:
          type: string
          description: Organization number of invoice issuer
          examples:
          - "5565027223"
        administration:
          type: string
          description: Administration
          examples:
          - Sundsvall Elnät
    Invoice:
      type: object
      description: Invoice model
      properties:
        customerNumber:
          type: string
          description: Customer number
          examples:
          - "39195"
          readOnly: true
        customerType:
          $ref: "#/components/schemas/CustomerType"
          readOnly: true
        facilityIds:
          type: array
          description: List of facility ids
          examples:
          - "735999109151404321"
          - "735999109151401234"
          items:
            type: string
          readOnly: true
          uniqueItems: true
        invoiceDescriptions:
          type: array
          description: List of descriptions
          examples:
          - Fjärrvärme
          - Elnät
          items:
            type: string
          readOnly: true
          uniqueItems: true
        invoiceNumber:
          type: integer
          format: int64
          description: Invoice number
          examples:
          - "767915994"
          readOnly: true
        invoiceDate:
          type: string
          format: date
          description: Invoice date
          examples:
          - 2022-01-15
          readOnly: true
        invoiceName:
          type: string
          description: invoice name
          examples:
          - 765801493.pdf
          readOnly: true
        invoiceType:
          type: string
          description: Invoice type
          examples:
          - Faktura
          readOnly: true
        invoiceStatus:
          type: string
          description: Invoice status
          examples:
          - Skickad
          readOnly: true
        ocrNumber:
          type: integer
          format: int64
          description: Ocr number
          examples:
          - "767915994"
          readOnly: true
        dueDate:
          type: string
          format: date
          description: Due date
          examples:
          - 2022-01-31
          readOnly: true
        totalAmount:
          type: number
          description: Total amount
          examples:
          - "1116.00"
          readOnly: true
        amountVatIncluded:
          type: number
          description: Amount included VAT
          examples:
          - "1115.62"
          readOnly: true
        amountVatExcluded:
          type: number
          description: Amount excluded VAT
          examples:
          - "892.50"
          readOnly: true
        vatEligibleAmount:
          type: number
          description: Amount eligible for VAT
          examples:
          - "892.50"
          readOnly: true
        rounding:
          type: number
          description: Rounding
          examples:
          - "0.38"
          readOnly: true
        vat:
          type: number
          description: VAT
          examples:
          - "892.50"
          readOnly: true
        reversedVat:
          type: boolean
          description: Reversed VAT
          examples:
          - "false"
          readOnly: true
        currency:
          type: string
          description: Currency
          examples:
          - sek
          readOnly: true
        organizationGroup:
          type: string
          description: Organization group
          examples:
          - stadsbacken
          readOnly: true
        organizationNumber:
          type: string
          description: Organization number of invoice issuer
          examples:
          - "5565027223"
          readOnly: true
        administration:
          type: string
          description: Adminstration
          examples:
          - Sundsvall Elnät
          readOnly: true
        street:
          type: string
          description: Street
          examples:
          - Storgatan 44
          readOnly: true
        postCode:
          type: string
          description: Postal code
          examples:
          - "85230"
          readOnly: true
        city:
          type: string
          description: City
          examples:
          - Sundsvall
          readOnly: true
        careOf:
          type: string
          description: Care of address
          examples:
          - Agatha Malm
          readOnly: true
        pdfAvailable:
          type: boolean
          description: Is pdf-version of invoice available
          examples:
          - "false"
          readOnly: true
    InvoiceResponse:
      type: object
      description: Invoice response model
      properties:
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
            readOnly: true
          readOnly: true
        _meta:
          $ref: "#/components/schemas/PagingAndSortingMetaData"
          readOnly: true
    PagingAndSortingMetaData:
      type: object
      description: PagingAndSortingMetaData model
      properties:
        page:
          type: integer
          format: int32
          description: Current page
          examples:
          - "5"
          readOnly: true
        limit:
          type: integer
          format: int32
          description: Displayed objects per page
          examples:
          - "20"
          readOnly: true
        count:
          type: integer
          format: int32
          description: Displayed objects on current page
          examples:
          - "13"
          readOnly: true
        totalRecords:
          type: integer
          format: int64
          description: Total amount of hits based on provided search parameters
          examples:
          - "98"
          readOnly: true
        totalPages:
          type: integer
          format: int32
          description: Total amount of pages based on provided search parameters
          examples:
          - "23"
          readOnly: true
        sortBy:
          type: array
          items:
            type: string
            description: The properties to sort by
            examples:
            - property
            readOnly: true
          readOnly: true
        sortDirection:
          $ref: "#/components/schemas/Direction"
          description: The sort order direction
          examples:
          - ASC
    InvoiceDetail:
      type: object
      description: Invoice detail model
      properties:
        invoiceNumber:
          type: integer
          format: int64
          description: Invoice number
          examples:
          - "767915994"
          readOnly: true
        amount:
          type: number
          description: Amount
          examples:
          - "66.97"
          readOnly: true
        amountVatExcluded:
          type: number
          description: Amount excluded VAT
          examples:
          - "53.57"
          readOnly: true
        vat:
          type: number
          description: VAT
          examples:
          - "13.40"
          readOnly: true
        vatRate:
          type: number
          format: double
          description: VAT rate
          examples:
          - "25.0"
          readOnly: true
        quantity:
          type: number
          format: double
          description: Quantity
          examples:
          - "154.39"
          readOnly: true
        unit:
          type: string
          description: Unit
          examples:
          - kWh
          readOnly: true
        unitPrice:
          type: number
          description: Price per unit
          examples:
          - "0.347"
          readOnly: true
        periodFrom:
          type: string
          description: Period from
          examples:
          - 2022-01-01
          readOnly: true
        periodTo:
          type: string
          description: Period to
          examples:
          - 2022-01-31
          readOnly: true
        description:
          type: string
          description: Description
          examples:
          - This is a description
          readOnly: true
        productCode:
          type: integer
          format: int32
          description: Product code
          examples:
          - "1404"
          readOnly: true
        productName:
          type: string
          description: Product name
          examples:
          - Elöverföring
          readOnly: true
        organizationNumber:
          type: string
          description: Organization number of invoice issuer
          examples:
          - "5565027223"
          readOnly: true
        administration:
          type: string
          description: Administration
          examples:
          - Name of the administration
          readOnly: true
        facilityId:
          type: string
          description: Facility id
          examples:
          - "735999109151401011"
          readOnly: true
...
    CustomerEngagement:
      type: object
      description: Customer engagement model
      properties:
        partyId:
          type: string
          description: PartyId (e.g. a personId or an organizationId)
          examples:
          - 81471222-5798-11e9-ae24-57fa13b361e1
          readOnly: true
        customerType:
          $ref: "#/components/schemas/CustomerType"
          readOnly: true
        customerNumber:
          type: string
          description: Customer number
          examples:
          - "10007"
          readOnly: true
        organizationNumber:
          type: string
          description: Organization number for counterpart of engagement
          examples:
          - "5565027223"
          readOnly: true
        organizationName:
          type: string
          description: Organization name for counterpart of engagement
          examples:
          - Sundsvall Elnät
          readOnly: true
        active:
          type: boolean
          description: "Indicates customer status, if not active then the moveInDate\
            \ holds information on when the customer will be activated"
          examples:
          - "true"
          readOnly: true
        moveInDate:
          type: string
          format: date
          description: The prospective customer's move-in date
          readOnly: true
    CustomerEngagementResponse:
      type: object
      description: Customer engagement response model
      properties:
        customerEngagements:
          type: array
          items:
            $ref: "#/components/schemas/CustomerEngagement"
            readOnly: true
          readOnly: true
        _meta:
          $ref: "#/components/schemas/PagingAndSortingMetaData"
          readOnly: true
...
    CustomerDetailsResponse:
      type: object
      description: Customer details response model
      properties:
        customerDetails:
          type: array
          items:
            $ref: "#/components/schemas/CustomerDetails"
            readOnly: true
          readOnly: true
        _meta:
          $ref: "#/components/schemas/PagingAndSortingMetaData"
          readOnly: true
...
  securitySchemes: {}